backend/requirements.txt
azure-functions
openai
azure-search-documents==11.4.0
azure-core
python-dotenv
